+++ packages/tailwindcss/src/utilities.ts
@@ -254,7 +254,7 @@ function resolveThemeColor<T extends ThemeKey>(
       break
     }
     case 'current': {
-      value = 'currentColor'
+      value = 'currentcolor'
       break
     }
     default: {
@@ -5448,7 +5448,7 @@ export function createUtilities(theme: Theme) {
 
     staticUtility('ring-inset', [boxShadowProperties, ['--tw-ring-inset', 'inset']])
 
-    let defaultRingColor = theme.get(['--default-ring-color']) ?? 'currentColor'
+    let defaultRingColor = theme.get(['--default-ring-color']) ?? 'currentcolor'
     function ringShadowValue(value: string) {
       return `var(--tw-ring-inset,) 0 0 0 calc(${value} + var(--tw-ring-offset-width)) var(--tw-ring-color, ${defaultRingColor})`
     }
@@ -5524,7 +5524,7 @@ export function createUtilities(theme: Theme) {
     ])
 
     function insetRingShadowValue(value: string) {
-      return `inset 0 0 0 ${value} var(--tw-inset-ring-color, currentColor)`
+      return `inset 0 0 0 ${value} var(--tw-inset-ring-color, currentcolor)`
     }
     utilities.functional('inset-ring', (candidate) => {
       if (!candidate.value) {
@@ -6192,13 +6192,20 @@ function alphaReplacedShadowProperties(
     return varInjector(replaceAlpha(color, alpha))
   })
 
+  function applyPrefix(x: string) {
+    if (!prefix) return x
+    return segment(x, ',')
+      .map((value) => prefix + value)
+      .join(',')
+  }
+
   if (requiresFallback) {
     return [
-      decl(property, prefix + replaceShadowColors(value, varInjector)),
-      rule('@supports (color: lab(from red l a b))', [decl(property, prefix + replacedValue)]),
+      decl(property, applyPrefix(replaceShadowColors(value, varInjector))),
+      rule('@supports (color: lab(from red l a b))', [decl(property, applyPrefix(replacedValue))]),
     ]
   } else {
-    return [decl(property, prefix + replacedValue)]
+    return [decl(property, applyPrefix(replacedValue))]
   }
 }